🩹 oups forgot to move a test into slow ones
